requirements.txt
# Is this how they work? I'm just manually typing this

pywin32
win32api
pyserial
